+++ icevision/models/torchvision_models/faster_rcnn/__init__.py
@@ -1,6 +1,7 @@
 from icevision.models.torchvision_models import backbones
 from icevision.models.torchvision_models.loss_fn import *
 
+from icevision.models.torchvision_models.faster_rcnn.parser import *
 from icevision.models.torchvision_models.faster_rcnn.dataloaders import *
 from icevision.models.torchvision_models.faster_rcnn.model import *
 from icevision.models.torchvision_models.faster_rcnn.prediction import *