+++ icevision/models/torchvision_models/keypoint_rcnn/parser.py
@@ -1,0 +1,5 @@
+__all__ = ["Parser", "DefaultParser"]
+
+from icevision.models.torchvision_models.parser import *
+
+DefaultParser = KeyPointsDefaultParser