+++ icevision/models/torchvision_models/mask_rcnn/__init__.py
@@ -2,6 +2,7 @@
 from icevision.models.torchvision_models import backbones
 from icevision.models.torchvision_models.loss_fn import *
 
+from icevision.models.torchvision_models.mask_rcnn.parser import *
 from icevision.models.torchvision_models.mask_rcnn.dataloaders import *
 from icevision.models.torchvision_models.mask_rcnn.model import *
 from icevision.models.torchvision_models.mask_rcnn.prediction import *