+++ icevision/models/torchvision_models/retinanet/__init__.py
@@ -1,6 +1,7 @@
 from icevision.models.torchvision_models.loss_fn import *
 
 from icevision.models.torchvision_models.retinanet import backbones
+from icevision.models.torchvision_models.retinanet.parser import *
 from icevision.models.torchvision_models.retinanet.dataloaders import *
 from icevision.models.torchvision_models.retinanet.model import *
 from icevision.models.torchvision_models.retinanet.prediction import *