+++ icevision/parsers/parser.py
@@ -37,6 +37,11 @@ class Parser(ImageidMixin, SizeMixin, ParserInterface, ABC):
     ```
     """
 
+    def __init__(self, class_map: Optional[ClassMap] = None):
+        self.class_map = class_map or ClassMap()
+        if class_map is None:
+            self.class_map.unlock()
+
     @abstractmethod
     def __iter__(self) -> Any:
         pass
@@ -89,6 +94,7 @@ class Parser(ImageidMixin, SizeMixin, ParserInterface, ABC):
     def parse(
         self,
         data_splitter: DataSplitter = None,
+        background_id: Optional[int] = None,
         idmap: IDMap = None,
         autofix: bool = True,
         show_pbar: bool = True,
@@ -116,6 +122,7 @@ class Parser(ImageidMixin, SizeMixin, ParserInterface, ABC):
             )
             return pickle.load(open(Path(cache_filepath), "rb"))
         else:
+            self.class_map.set_background(background_id)
             idmap = idmap or IDMap()
             data_splitter = data_splitter or RandomSplitter([0.8, 0.2])
             records = self.parse_dicted(show_pbar=show_pbar, idmap=idmap)